# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/team": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"properties": {
													"linkedin": {
														"type": "string"
													},
													"github": {
														"type": "string"
													},
													"email": {
														"type": "string"
													},
													"createdAt": {
														"type": "string",
														"format": "date-time"
													},
													"image": {
														"type": "string"
													},
													"bio": {
														"type": "string"
													},
													"role": {
														"type": "string"
													},
													"name": {
														"type": "string"
													},
													"id": {
														"type": "number",
														"format": "double"
													}
												},
												"required": [
													"linkedin",
													"github",
													"email",
													"createdAt",
													"image",
													"bio",
													"role",
													"name",
													"id"
												],
												"type": "object"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"error",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"member": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"member": {
													"properties": {
														"linkedin": {
															"type": "string"
														},
														"github": {
															"type": "string"
														},
														"email": {
															"type": "string"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"image": {
															"type": "string"
														},
														"bio": {
															"type": "string"
														},
														"role": {
															"type": "string"
														},
														"name": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"linkedin",
														"github",
														"email",
														"createdAt",
														"image",
														"bio",
														"role",
														"name",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"member",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"member": {},
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"error",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"github": {
										"type": "string"
									},
									"linkedin": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"image": {
										"type": "string"
									},
									"bio": {
										"type": "string"
									},
									"role": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"role",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/team/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"linkedin": {
													"type": "string"
												},
												"github": {
													"type": "string"
												},
												"email": {
													"type": "string"
												},
												"createdAt": {
													"type": "string",
													"format": "date-time"
												},
												"image": {
													"type": "string"
												},
												"bio": {
													"type": "string"
												},
												"role": {
													"type": "string"
												},
												"name": {
													"type": "string"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"linkedin",
												"github",
												"email",
												"createdAt",
												"image",
												"bio",
												"role",
												"name",
												"id"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"error",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"member": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"member": {
													"properties": {
														"linkedin": {
															"type": "string"
														},
														"github": {
															"type": "string"
														},
														"email": {
															"type": "string"
														},
														"createdAt": {
															"type": "string",
															"format": "date-time"
														},
														"image": {
															"type": "string"
														},
														"bio": {
															"type": "string"
														},
														"role": {
															"type": "string"
														},
														"name": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														}
													},
													"required": [
														"linkedin",
														"github",
														"email",
														"createdAt",
														"image",
														"bio",
														"role",
														"name",
														"id"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"member",
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"member": {},
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"error",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"github": {
										"type": "string"
									},
									"linkedin": {
										"type": "string"
									},
									"email": {
										"type": "string"
									},
									"image": {
										"type": "string"
									},
									"bio": {
										"type": "string"
									},
									"role": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												}
											},
											"required": [
												"error",
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Team"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}